fix(ci): green main — product-ui tooltip assertions + test_cloud_api length
Two fixes for the only red jobs on main's CI, which gate Deploy Staging:

1. product-ui CloudChatComposer copy-feedback tests: #811 replaced the
   hand-rolled @proliferate/ui Tooltip (opened on mouseenter, rendered
   role="tooltip" inline) with a Radix-backed kit seam. Radix tooltips
   open on pointermove/focus and portal role="tooltip" only while open,
   so the mouseenter-driven assertions never saw the tooltip. The copied
   feedback itself still works (aria-label swap, sr-only announcement,
   check icon, tooltip content flips to "Copied"), so update the tests
   to open the tooltip via focus/blur — the deterministic path under
   jsdom that avoids Radix's pointer delay and grace-area tracking.

2. server/tests/integration/test_cloud_api.py went to 687 lines
   (allowlisted 679) when #846 added GitHub-App repo-authority seeding.
   Trim instead of bumping: extract the shared seeding/setup helpers
   (register_and_login, link_github_account,
   seed_github_app_repo_authority, create_organization_for_user,
   add_organization_member) into tests/integration/cloud_api_helpers.py
   (billing_accounting_helpers pattern) and delete four dead helpers
   left behind by #846's test sweep (_billing_subject_for_user,
   _link_secondary_account, _list_mcp_connections,
   _list_mcp_connection_auths). File is now 478 lines, under the 600
   base max, so drop its allowlist entry; also tighten the
   test_auth_flow.py entry to 2052 (#846 trimmed it by 5 lines, and the
   checker errors on stale-too-generous entries).
