+++ scripts/max_lines_allowlist.txt
@@ -41,8 +41,7 @@ server/proliferate/server/cloud/agent_auth/models.py 563 managed sandbox desktop
 server/proliferate/server/cloud/agent_auth/runtime_keys.py 671 existing cloud agent auth runtime key service pending gateway/router split
 server/proliferate/server/cloud/secrets/api.py 416 existing cloud secrets API pending route split
 server/proliferate/constants/cloud.py 660 existing cloud constants module with command and target config enums
-server/tests/integration/test_auth_flow.py 2057 existing server oversized integration test
+server/tests/integration/test_auth_flow.py 2052 existing server oversized integration test
 server/tests/integration/test_billing_accounting.py 872 existing server oversized integration test
 server/tests/integration/test_billing_api.py 1918 existing server oversized integration test
-server/tests/integration/test_cloud_api.py 679 existing server oversized integration test
 server/tests/integration/test_stripe_webhooks.py 1456 existing server oversized integration test
